Trait pallet_remarks::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type RemarkDispatchHandler: RemarkDispatchHandler<RemarkArgs<Self>>;
    type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + GetDispatchInfo + From<Call<Self>> + IsSubType<Call<Self>> + IsType<<Self as Config>::RuntimeCall>;
    type WeightInfo: WeightInfo;
    type Remark: Parameter + Member + Default;
    type MaxRemarksPerCall: Get<u32>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type RemarkDispatchHandler: RemarkDispatchHandler<RemarkArgs<Self>>

The handler used to check remarks before and after call dispatch.

type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + GetDispatchInfo + From<Call<Self>> + IsSubType<Call<Self>> + IsType<<Self as Config>::RuntimeCall>

The overarching call type.

type WeightInfo: WeightInfo

Weight information.

type Remark: Parameter + Member + Default

The type attached to the remark event.

type MaxRemarksPerCall: Get<u32>

Type that restrains the maximum remarks that can be attached to a call.

Object Safety

This trait is not object safe.